+++ app/Main.hs
@@ -15,14 +15,15 @@ import Data.Time (Day, UTCTime (..))
 import Distribution.Client.Config (loadConfig, savedGlobalFlags)
 import Distribution.Client.GlobalFlags (globalCacheDir)
 import Distribution.Compat.NonEmptySet (singleton)
+import Distribution.Parsec (simpleParsec)
 import Distribution.Pretty (pretty)
 import Distribution.Simple.Flag (fromFlag)
 import Distribution.Types.Dependency (Dependency (..))
 import Distribution.Types.LibraryName (LibraryName (..))
 import Distribution.Types.PackageName (PackageName, mkPackageName, unPackageName)
 import Distribution.Types.VersionRange (VersionRange)
-import Hackage.RevDeps (getReverseDependencies, getTransitiveReverseDependencies)
-import Options.Applicative (Parser, ReadM, auto, execParser, fullDesc, help, helper, info, long, metavar, option, optional, progDesc, strArgument, switch)
+import Hackage.RevDeps (ExtractDependenciesMode (..), getReverseDependencies, getTransitiveReverseDependencies)
+import Options.Applicative (Parser, ReadM, argument, auto, execParser, fullDesc, help, helper, info, long, maybeReader, metavar, option, optional, progDesc, switch)
 import Options.Applicative.NonEmpty (some1)
 import System.Console.ANSI (hSupportsANSI, hyperlinkCode)
 import System.FilePath ((</>))
@@ -43,13 +44,13 @@ parseArgs = do
           <> help "Timestamp of index state at which to stop scanning, YYYY-MM-DD"
   cnfPackageNames <-
     some1 $
-      strArgument $
+      argument (maybeReader simpleParsec) $
         metavar "PKGS"
           <> help "Package names to scan Hackage for their reverse dependencies"
   cnfTransitive <-
     switch $
       long "transitive"
-        <> help "Count transitive (both direct and indirect) dependencies. This mode is imprecise when passing multiple package names at once."
+        <> help "Count transitive (both direct and indirect) dependencies. This mode is imprecise when passing multiple package names at once. Also in this mode we follow only dependencies of the main library"
   pure Config {..}
 
 main :: IO ()
@@ -63,7 +64,7 @@ main = do
   cnf <- loadConfig minBound mempty
   let cacheDir = fromFlag $ globalCacheDir $ savedGlobalFlags cnf
       idx = cacheDir </> hackageHaskellOrg </> "01-index.tar"
-      func = if cnfTransitive then getTransitiveReverseDependencies else getReverseDependencies
+      func = if cnfTransitive then getTransitiveReverseDependencies OnlyMainLib else getReverseDependencies AllComponents
   pkgs <- func args idx cnfIndexState
   report args $ M.delete (mkPackageName "acme-everything") pkgs